[home][s]: change info economy copy to use that from the /information-economy/ page - cf #49.
Text I removed was:

How can we can build an open information age? And why should we do it?
Will robots take all the jobs - and should we care? How to preserve
freedom in a world of Googles and Facebooks who have the power to shape
how we think and act? How can we harness the full power of digital tech
for innovation and creativity? Concerned about growing inequality and
the digital divide? Want to build an economy and society fit for the
information age? Open information is the biggest policy opportunity of
the 21st century with answers for all of these questions. Openness is
central not only to creating a more innovative and transparent society
but to creating one which is fair, free, healthy and wealthy.