[xs][refactor] get rid of directores in post folder ahead of slug correction.
Also better for linking in obsidian etc.

There are a few posts in `/post` folder that are in directories with markdown file named `index.md` e.g. /post/zettelkasten/index. Move these markdown files up a directory renaming the index.md to be named like directory it was in with .md
  extension. After that move any assets that rename in the directories / subfolders to the images/ folder and then correct any links to the renamed markdown files in first step to the images.