[m][refactor]: Rename files in `/post` directory to be consistent with old hugo generated urls.
## Situation

Some links on my website are broken because page urls have changed. For example:

Old: https://rufuspollock.com/2016/11/26/fake-news-post-truth-is-it-news-and-what-can-we-do/

Current: https://rufuspollock.com/post/2016-11-26-fake-news-post-truth-is-it-new

Another example:

Old: https://rufuspollock.com/2020/07/01/zettelkasten-method/

Current: https://rufuspollock.com/post/zettelkasten-method/

### Diagnosis

The hugo website builder generated url paths with a specific algorithm (described below). Meanwhile, uses simple file name based routing i.e. the url is the file path on disk. These are different!

Hugo generates urls for posts in a form like this:

```
/yyyy/mm/dd/slug
```

Where:

- The original filename is not used at all
- The `/yyyy/mm/dd` is derived from the `date` frontmatter field
- The `slug` for the post is created via this algorithm:
    - Use the `slug` frontmatter field in the file if it exists
    - Otherwise use the `title` field in the frontmatter. Convert it ot a slug by: a) lowercase b) removing all non alphanumeric characters like brackets, question marks etc c) remove `the` or `a` if it starts the title d) if `'s` attach `s` to previous word e) kebab-case

### Prognosis

I can fix the `/yyyy/mm/dd/{slug}`  url routing using a redirect in cloudflare to redirect to `/post/{slug}` and i've done that.

What i need to fix now is to update local file names on disk in `/post` directory to have their file name be the slug.

Two steps:

- Compute `slug` based on hugo algorithm described ablove
- Update the file to that name if different

Please prepare and apply a script to do this in python or node. The script should be written to sandbox directory.

NB: I only want this applied to files in  the `/post` folder.